refactor styling files with related pages or components
