+++ client/src/components/ItemSlider.css
@@ -16,7 +16,8 @@
   gap: 1rem;
   overflow-x: hidden;
   scroll-behavior: smooth;
-  width: 640px;
+  width: 100%;
+  max-width: 850px;
   padding: 1rem 0;
 }
