+++ client/src/components/Pagination.css
@@ -1,0 +1,42 @@
+.pagination {
+  text-align: center;
+  margin-top: 1.5rem;
+}
+
+.pagination-btn {
+  background-color: #f0f0f0;
+  border: 1px solid #ccc;
+  padding: 0.5rem 1.2rem;
+  margin: 0 6px;
+  cursor: pointer;
+  border-radius: 6px;
+  font-size: 1rem;
+  font-weight: 500;
+  color: #444;
+  transition:
+    background-color 0.3s ease,
+    color 0.3s ease,
+    border-color 0.3s ease;
+  user-select: none;
+}
+
+.pagination-btn:hover:not(:disabled) {
+  background-color: #0798e6;
+  color: white;
+  border-color: #006bb3;
+}
+
+.pagination-btn:disabled {
+  cursor: not-allowed;
+  opacity: 0.5;
+  background-color: #ddd;
+  color: #999;
+}
+
+.pagination-btn.active {
+  font-weight: 700;
+  background-color: #0798e6;
+  border-color: #006bb3;
+  color: white;
+  box-shadow: 0 0 8px rgba(7, 152, 230, 0.7);
+}
