+++ client/src/pages/Home/Home.jsx
@@ -4,7 +4,7 @@ import Header from "../../components/Header";
 import Footer from "../../components/Footer";
 import ItemSlider from "../../components/ItemSlider";
 import wallpaper from "../../assets/wallpaper.jpg";
-import "../../styles/HomeStyle.css";
+import "./Home.css";
 
 const Home = () => {
   const [searchItem, setSearchItem] = useState("");
